Extract board hooks - useBoardStyle - useFloatingActions - useFloatingFocus
