+++ packages/scrabble-solver/src/components/Board/hooks/index.ts
@@ -1,2 +1,5 @@
 export { default as useBackgroundImage } from './useBackgroundImage';
+export { default as useBoardStyle } from './useBoardStyle';
+export { default as useFloatingActions } from './useFloatingActions';
+export { default as useFloatingFocus } from './useFloatingFocus';
 export { default as useGrid } from './useGrid';
