+++ packages/scrabble-solver/src/components/Board/hooks/useFloatingFocus.ts
@@ -1,0 +1,10 @@
+import { autoUpdate, useFloating } from '@floating-ui/react';
+
+const useFloatingFocus = () => {
+  return useFloating({
+    placement: 'top-start',
+    whileElementsMounted: autoUpdate,
+  });
+};
+
+export default useFloatingFocus;
